+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQSPHKGBK54RPYFHFHWHY5NT"
+        "01KQSSTN56MJFT59CRVNTM5M83"
     )
     r"""The unique identifier of the data source"""