+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQSPHKGA5NZDXFY61BYARDXT"
+        "01KQSSTN55GDVW1V74DQ93P6ST"
     )
     r"""The unique identifier of the data source"""